separate JwtAccess class into own file
