+++ src/ServiceAccountCredentials.php
@@ -133,73 +133,3 @@ class ServiceAccountCredentials extends CredentialsLoader
     return $jwtCreds->updateMetadata($metadata, $authUri, $client);
   }
 }
-
-/**
- * Authenticates requests using Google's Service Account credentials via
- * JWT Access.
- *
- * This class allows authorizing requests for service accounts directly
- * from credentials from a json key file downloaded from the developer
- * console (via 'Generate new Json Key').  It is not part of any OAuth2
- * flow, rather it creates a JWT and sends that as a credential.
- */
-class ServiceAccountJwtAccessCredentials extends CredentialsLoader
-{
-  /**
-   * Create a new ServiceAccountJwtAccessCredentials.
-   *
-   * @param array jsonKey JSON credentials.
-   */
-  public function __construct($jsonKey)
-  {
-    if (!array_key_exists('client_email', $jsonKey)) {
-      throw new \InvalidArgumentException(
-          'json key is missing the client_email field');
-    }
-    if (!array_key_exists('private_key', $jsonKey)) {
-      throw new \InvalidArgumentException(
-          'json key is missing the private_key field');
-    }
-    $this->auth = new OAuth2([
-      'issuer' => $jsonKey['client_email'],
-      'sub' => $jsonKey['client_email'],
-      'signingAlgorithm' => 'RS256',
-      'signingKey' => $jsonKey['private_key'],
-    ]);
-  }
-
-  /**
-   * Updates metadata with the authorization token
-   *
-   * @param $metadata array metadata hashmap
-   * @param $authUri string optional auth uri
-   * @param $client optional client interface
-   *
-   * @return array updated metadata hashmap
-   */
-  public function updateMetadata($metadata,
-                                 $authUri = null,
-                                 ClientInterface $client = null)
-  {
-    if (empty($authUri)) {
-      return $metadata;
-    }
-
-    $this->auth->setAudience($authUri);
-    return parent::updateMetadata($metadata, $authUri, $client);
-  }
-
- /**
-  * Implements FetchAuthTokenInterface#fetchAuthToken.
-  */
-  public function fetchAuthToken(ClientInterface $unusedClient = null)
-  {
-    $audience = $this->auth->getAudience();
-    if (empty($audience)) {
-      return null;
-    }
-
-    $access_token = $this->auth->toJwt();
-    return array('access_token' => $access_token);
-  }
-}
